forsättning uppgift.

+++ LektionAnas20231031/Doctor.cs
@@ -9,7 +9,7 @@ namespace LektionAnas20231031
     internal class Doctor : Person
     {
         public int Salary { get; set; }
-        public int Workhours { get; set; }
+        public double Workhours { get; set; }
 
         public static void WritePrescription()
         {
